% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_vector}
\alias{chunk_vector}
\title{Split a vector into equally sized chunks}
\usage{
chunk_vector(x, size)
}
\arguments{
\item{x}{Vector to split.}

\item{size}{Maximum chunk size.}
}
\value{
A list of vector chunks (possibly empty).
}
\description{
Convenience wrapper used to batch large vectors in API requests.
}
