% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_helpers.R
\name{cr_fetch_works}
\alias{cr_fetch_works}
\title{Fetch Crossref works}
\usage{
cr_fetch_works(filters = list(), limit = 1000, ...)
}
\arguments{
\item{filters}{A named list of filters passed to \code{\link[rcrossref:cr_works]{rcrossref::cr_works()}}. Use
Crossref filter keys such as \code{"query.affiliation"} or \code{"funder"}.}

\item{limit}{Number of records to request. Defaults to 1000.}

\item{...}{Additional arguments forwarded to \code{\link[rcrossref:cr_works]{rcrossref::cr_works()}}.}
}
\value{
A tibble of Crossref works (possibly empty).
}
\description{
Fetch Crossref works
}
