% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref_helpers.R
\name{cr_filter_affiliations}
\alias{cr_filter_affiliations}
\title{Filter Crossref works by affiliation pattern}
\usage{
cr_filter_affiliations(
  works,
  pattern,
  column = "author",
  matcher = cr_match_author_affiliation,
  ...
)
}
\arguments{
\item{works}{Tibble from \code{\link[=cr_fetch_works]{cr_fetch_works()}}.}

\item{pattern}{Pattern to match within the \code{column}.}

\item{column}{Name of the column containing author information. Defaults to
\code{"author"}.}

\item{matcher}{Function used to detect matches. Defaults to
\code{\link[=cr_match_author_affiliation]{cr_match_author_affiliation()}}.}

\item{...}{Additional arguments forwarded to the matcher.}
}
\value{
Filtered tibble.
}
\description{
Filter Crossref works by affiliation pattern
}
