% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_fetch.R
\name{fetch_crossref_metadata}
\alias{fetch_crossref_metadata}
\title{Fetch Crossref metadata for a set of DOIs}
\usage{
fetch_crossref_metadata(
  dois,
  chunk_size = 50,
  pause = 0.1,
  progress = interactive()
)
}
\arguments{
\item{dois}{Character vector of DOI strings.}

\item{chunk_size}{Number of DOIs to include per request batch.}

\item{pause}{Seconds to wait between requests, helping to respect rate
limits.}

\item{progress}{Logical; show a text progress bar? Defaults to
\code{interactive()}.}
}
\value{
Tibble of Crossref works (possibly empty).
}
\description{
Fetch Crossref metadata for a set of DOIs
}
