% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_fetch.R
\name{fetch_openalex_metadata}
\alias{fetch_openalex_metadata}
\title{Fetch OpenAlex metadata for a set of DOIs}
\usage{
fetch_openalex_metadata(dois, pause = 0.1, progress = interactive())
}
\arguments{
\item{dois}{Character vector of DOI strings.}

\item{pause}{Seconds to wait between requests.}

\item{progress}{Logical; show a text progress bar? Defaults to
\code{interactive()}.}
}
\value{
Tibble of OpenAlex works (possibly empty).
}
\description{
Fetch OpenAlex metadata for a set of DOIs
}
